[bugfix] ScrollWidget, need to cast to int

+++ labelMatch/widgets/scrollwidget.py
@@ -19,4 +19,4 @@ class ScrollWidget(QScrollArea):
     def scroll_request(self, delta, orientation):
         units = - delta / (8 * 15)
         bar = self.scrollbars[orientation]
-        bar.setValue(bar.value() + bar.singleStep() * units)
+        bar.setValue(int(bar.value() + bar.singleStep() * units))
